app/requirements.txt
# Core
fastapi
uvicorn
cassandra-driver
pydantic
pydantic-settings

# Redis + Rate limiting
redis
slowapi

# GeoIP
geoip2

# Env
python-dotenv

# Testing
pytest
httpx
